refactor: renamed internal sources' directory
