+++ src/comparison/isEq.ts
@@ -1,4 +1,4 @@
-import { tagOf, TypeTag } from "../.core/tag";
+import { tagOf, TypeTag } from "../.internal/tag";
 import { isFunc } from "../type/isFunc";
 import { isStrictEq } from "./isStrictEq";
