+++ src/type/isAsyncGenFunc.ts
@@ -1,4 +1,4 @@
-import { tagOf, TypeTag } from "../.core/tag";
+import { tagOf, TypeTag } from "../.internal/tag";
 
 /**
  * Checks whether value is an AsyncGeneratorFunction object.
